refactor: compentize UseCaseSection
+ remove unused imports
